change mouth initial position to closed
