+++ tsconfig.json
@@ -22,6 +22,6 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "components/FindShapeIndex.js", "__tests__/api/transcribe/transcribe.api.spec.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "components/FindShapeIndex.js", "__tests__/api/transcribe/transcribe.api.spec.ts", "hooks/useTTS.ts"],
   "exclude": ["node_modules"]
 }
